Refactor: Clean up Overlord guice dependencies
