+++ indexing-service/src/main/java/org/apache/druid/indexing/overlord/TaskMaster.java
@@ -19,254 +19,61 @@
 
 package org.apache.druid.indexing.overlord;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.base.Optional;
 import com.google.inject.Inject;
-import org.apache.druid.client.indexing.IndexingService;
-import org.apache.druid.curator.discovery.ServiceAnnouncer;
-import org.apache.druid.discovery.DruidLeaderSelector;
-import org.apache.druid.discovery.DruidLeaderSelector.Listener;
-import org.apache.druid.guice.annotations.Self;
-import org.apache.druid.indexing.common.actions.SegmentAllocationQueue;
 import org.apache.druid.indexing.common.actions.TaskActionClient;
 import org.apache.druid.indexing.common.actions.TaskActionClientFactory;
 import org.apache.druid.indexing.common.task.Task;
-import org.apache.druid.indexing.common.task.TaskContextEnricher;
 import org.apache.druid.indexing.overlord.autoscaling.ScalingStats;
-import org.apache.druid.indexing.overlord.config.DefaultTaskConfig;
-import org.apache.druid.indexing.overlord.config.TaskLockConfig;
-import org.apache.druid.indexing.overlord.config.TaskQueueConfig;
-import org.apache.druid.indexing.overlord.duty.OverlordDutyExecutor;
 import org.apache.druid.indexing.overlord.supervisor.SupervisorManager;
-import org.apache.druid.java.util.common.lifecycle.Lifecycle;
-import org.apache.druid.java.util.common.lifecycle.LifecycleStart;
-import org.apache.druid.java.util.common.lifecycle.LifecycleStop;
-import org.apache.druid.java.util.emitter.EmittingLogger;
-import org.apache.druid.java.util.emitter.service.ServiceEmitter;
-import org.apache.druid.server.DruidNode;
-import org.apache.druid.server.coordinator.CoordinatorOverlordServiceConfig;
 import org.apache.druid.server.coordinator.stats.CoordinatorRunStats;
 import org.apache.druid.server.metrics.TaskCountStatsProvider;
 import org.apache.druid.server.metrics.TaskSlotCountStatsProvider;
 
 import javax.annotation.Nullable;
 import java.util.Map;
-import java.util.concurrent.atomic.AtomicReference;
-import java.util.concurrent.locks.ReentrantLock;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 /**
- * Encapsulates the indexer leadership lifecycle.
+ * Encapsulates various Overlord classes that allow querying and updating the
+ * current state of the Overlord leader.
  */
 public class TaskMaster implements TaskCountStatsProvider, TaskSlotCountStatsProvider
 {
-  private static final EmittingLogger log = new EmittingLogger(TaskMaster.class);
-
-  private final DruidLeaderSelector overlordLeaderSelector;
-  private final DruidLeaderSelector.Listener leadershipListener;
-
-  private final ReentrantLock giant = new ReentrantLock(true);
   private final TaskActionClientFactory taskActionClientFactory;
   private final SupervisorManager supervisorManager;
-
-  private final AtomicReference<Lifecycle> leaderLifecycleRef = new AtomicReference<>(null);
-
   private volatile TaskRunner taskRunner;
   private volatile TaskQueue taskQueue;
 
-  /**
-   * This flag indicates that all services has been started and should be true before calling
-   * {@link ServiceAnnouncer#announce}. This is set to false immediately once {@link Listener#stopBeingLeader()} is
-   * called.
-   */
-  private volatile boolean initialized;
+  private final AtomicBoolean isLeader = new AtomicBoolean(false);
 
   @Inject
   public TaskMaster(
-      final TaskLockConfig taskLockConfig,
-      final TaskQueueConfig taskQueueConfig,
-      final DefaultTaskConfig defaultTaskConfig,
-      final TaskLockbox taskLockbox,
-      final TaskStorage taskStorage,
-      final TaskActionClientFactory taskActionClientFactory,
-      @Self final DruidNode selfNode,
-      final TaskRunnerFactory runnerFactory,
-      final ServiceAnnouncer serviceAnnouncer,
-      final CoordinatorOverlordServiceConfig coordinatorOverlordServiceConfig,
-      final ServiceEmitter emitter,
-      final SupervisorManager supervisorManager,
-      final OverlordDutyExecutor overlordDutyExecutor,
-      @IndexingService final DruidLeaderSelector overlordLeaderSelector,
-      final SegmentAllocationQueue segmentAllocationQueue,
-      final ObjectMapper mapper,
-      final TaskContextEnricher taskContextEnricher
+      TaskActionClientFactory taskActionClientFactory,
+      SupervisorManager supervisorManager
   )
   {
-    this.supervisorManager = supervisorManager;
     this.taskActionClientFactory = taskActionClientFactory;
-
-    this.overlordLeaderSelector = overlordLeaderSelector;
-
-    final DruidNode node = coordinatorOverlordServiceConfig.getOverlordService() == null ? selfNode :
-                           selfNode.withService(coordinatorOverlordServiceConfig.getOverlordService());
-
-    this.leadershipListener = new DruidLeaderSelector.Listener()
-    {
-      @Override
-      public void becomeLeader()
-      {
-        giant.lock();
-
-        // I AM THE MASTER OF THE UNIVERSE.
-        log.info("By the power of Grayskull, I have the power!");
-
-        try {
-          taskRunner = runnerFactory.build();
-          taskQueue = new TaskQueue(
-              taskLockConfig,
-              taskQueueConfig,
-              defaultTaskConfig,
-              taskStorage,
-              taskRunner,
-              taskActionClientFactory,
-              taskLockbox,
-              emitter,
-              mapper,
-              taskContextEnricher
-          );
-
-          // Sensible order to start stuff:
-          final Lifecycle leaderLifecycle = new Lifecycle("task-master");
-          if (leaderLifecycleRef.getAndSet(leaderLifecycle) != null) {
-            log.makeAlert("TaskMaster set a new Lifecycle without the old one being cleared!  Race condition")
-               .emit();
-          }
-
-          leaderLifecycle.addManagedInstance(taskRunner);
-          leaderLifecycle.addManagedInstance(taskQueue);
-          leaderLifecycle.addManagedInstance(supervisorManager);
-          leaderLifecycle.addManagedInstance(overlordDutyExecutor);
-          leaderLifecycle.addHandler(
-              new Lifecycle.Handler()
-              {
-                @Override
-                public void start()
-                {
-                  segmentAllocationQueue.becomeLeader();
-                }
-
-                @Override
-                public void stop()
-                {
-                  segmentAllocationQueue.stopBeingLeader();
-                }
-              }
-          );
-
-          leaderLifecycle.addHandler(
-              new Lifecycle.Handler()
-              {
-                @Override
-                public void start()
-                {
-                  initialized = true;
-                  serviceAnnouncer.announce(node);
-                }
-
-                @Override
-                public void stop()
-                {
-                  serviceAnnouncer.unannounce(node);
-                }
-              }
-          );
-
-          leaderLifecycle.start();
-        }
-        catch (Exception e) {
-          throw new RuntimeException(e);
-        }
-        finally {
-          giant.unlock();
-        }
-      }
-
-      @Override
-      public void stopBeingLeader()
-      {
-        giant.lock();
-        try {
-          initialized = false;
-          final Lifecycle leaderLifecycle = leaderLifecycleRef.getAndSet(null);
-
-          if (leaderLifecycle != null) {
-            leaderLifecycle.stop();
-          }
-        }
-        finally {
-          giant.unlock();
-        }
-      }
-    };
+    this.supervisorManager = supervisorManager;
   }
 
-  /**
-   * Starts waiting for leadership. Should only be called once throughout the life of the program.
-   */
-  @LifecycleStart
-  public void start()
+  public void becomeLeader(TaskRunner taskRunner, TaskQueue taskQueue)
   {
-    giant.lock();
-
-    try {
-      overlordLeaderSelector.registerListener(leadershipListener);
-    }
-    finally {
-      giant.unlock();
-    }
+    this.taskRunner = taskRunner;
+    this.taskQueue = taskQueue;
+    isLeader.set(true);
   }
 
-  /**
-   * Stops forever (not just this particular leadership session). Should only be called once throughout the life of
-   * the program.
-   */
-  @LifecycleStop
-  public void stop()
+  public void stopBeingLeader()
   {
-    giant.lock();
-
-    try {
-      gracefulStopLeaderLifecycle();
-      overlordLeaderSelector.unregisterListener();
-    }
-    finally {
-      giant.unlock();
-    }
+    isLeader.set(false);
+    this.taskQueue = null;
+    this.taskRunner = null;
   }
 
-  /**
-   * Returns true if it's the leader and all its services have been initialized.
-   */
-  public boolean isLeader()
+  private boolean isLeader()
   {
-    return overlordLeaderSelector.isLeader() && initialized;
-  }
-
-  public String getCurrentLeader()
-  {
-    return overlordLeaderSelector.getCurrentLeader();
-  }
-
-  public Optional<String> getRedirectLocation()
-  {
-    String leader = overlordLeaderSelector.getCurrentLeader();
-    // do not redirect when
-    // leader is not elected
-    // leader is the current node
-    if (leader == null || leader.isEmpty() || overlordLeaderSelector.isLeader()) {
-      return Optional.absent();
-    } else {
-      return Optional.of(leader);
-    }
+    return isLeader.get();
   }
 
   public Optional<TaskRunner> getTaskRunner()
@@ -377,18 +184,6 @@ public class TaskMaster implements TaskCountStatsProvider, TaskSlotCountStatsPro
       return taskQueue.get().getQueueStats();
     } else {
       return CoordinatorRunStats.empty();
-    }
-  }
-
-  private void gracefulStopLeaderLifecycle()
-  {
-    try {
-      if (isLeader()) {
-        leadershipListener.stopBeingLeader();
-      }
-    }
-    catch (Exception ex) {
-      // fail silently since we are stopping anyway
     }
   }
