+++ indexing-service/src/main/java/org/apache/druid/indexing/overlord/http/OverlordResource.java
@@ -33,6 +33,7 @@ import org.apache.druid.audit.AuditEntry;
 import org.apache.druid.audit.AuditManager;
 import org.apache.druid.client.indexing.ClientTaskQuery;
 import org.apache.druid.common.config.ConfigManager.SetResult;
+import org.apache.druid.common.config.Configs;
 import org.apache.druid.common.config.JacksonConfigManager;
 import org.apache.druid.error.DruidException;
 import org.apache.druid.indexer.RunnerTaskState;
@@ -44,31 +45,23 @@ import org.apache.druid.indexer.TaskStatusPlus;
 import org.apache.druid.indexing.common.actions.TaskActionClient;
 import org.apache.druid.indexing.common.actions.TaskActionHolder;
 import org.apache.druid.indexing.common.task.Task;
-import org.apache.druid.indexing.overlord.ImmutableWorkerInfo;
+import org.apache.druid.indexing.overlord.DruidOverlord;
 import org.apache.druid.indexing.overlord.IndexerMetadataStorageAdapter;
 import org.apache.druid.indexing.overlord.TaskMaster;
 import org.apache.druid.indexing.overlord.TaskQueryTool;
-import org.apache.druid.indexing.overlord.TaskQueue;
 import org.apache.druid.indexing.overlord.TaskRunner;
 import org.apache.druid.indexing.overlord.TaskRunnerWorkItem;
 import org.apache.druid.indexing.overlord.WorkerTaskRunner;
 import org.apache.druid.indexing.overlord.WorkerTaskRunnerQueryAdapter;
-import org.apache.druid.indexing.overlord.autoscaling.ProvisioningStrategy;
 import org.apache.druid.indexing.overlord.autoscaling.ScalingStats;
 import org.apache.druid.indexing.overlord.http.security.TaskResourceFilter;
-import org.apache.druid.indexing.overlord.setup.DefaultWorkerBehaviorConfig;
 import org.apache.druid.indexing.overlord.setup.WorkerBehaviorConfig;
 import org.apache.druid.java.util.common.DateTimes;
-import org.apache.druid.java.util.common.IAE;
 import org.apache.druid.java.util.common.Intervals;
 import org.apache.druid.java.util.common.StringUtils;
 import org.apache.druid.java.util.common.UOE;
 import org.apache.druid.java.util.common.logger.Logger;
 import org.apache.druid.metadata.LockFilterPolicy;
-import org.apache.druid.metadata.TaskLookup;
-import org.apache.druid.metadata.TaskLookup.ActiveTaskLookup;
-import org.apache.druid.metadata.TaskLookup.CompleteTaskLookup;
-import org.apache.druid.metadata.TaskLookup.TaskLookupType;
 import org.apache.druid.server.http.HttpMediaType;
 import org.apache.druid.server.http.ServletResourceUtils;
 import org.apache.druid.server.http.security.ConfigResourceFilter;
@@ -84,7 +77,7 @@ import org.apache.druid.server.security.Resource;
 import org.apache.druid.server.security.ResourceAction;
 import org.apache.druid.server.security.ResourceType;
 import org.apache.druid.tasklogs.TaskLogStreamer;
-import org.joda.time.Duration;
+import org.apache.druid.utils.CollectionUtils;
 import org.joda.time.Interval;
 
 import javax.annotation.Nullable;
@@ -104,17 +97,12 @@ import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.Status;
 import java.io.InputStream;
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.concurrent.atomic.AtomicReference;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 /**
  *
@@ -125,6 +113,7 @@ public class OverlordResource
   private static final Logger log = new Logger(OverlordResource.class);
 
   private final TaskMaster taskMaster;
+  private final DruidOverlord overlord;
   private final TaskQueryTool taskQueryTool;
   private final IndexerMetadataStorageAdapter indexerMetadataStorageAdapter;
   private final TaskLogStreamer taskLogStreamer;
@@ -132,35 +121,16 @@ public class OverlordResource
   private final AuditManager auditManager;
   private final AuthorizerMapper authorizerMapper;
   private final WorkerTaskRunnerQueryAdapter workerTaskRunnerQueryAdapter;
-  private final ProvisioningStrategy provisioningStrategy;
 
   private final AuthConfig authConfig;
 
-  private AtomicReference<WorkerBehaviorConfig> workerConfigRef = null;
   private static final List<String> API_TASK_STATES = ImmutableList.of("pending", "waiting", "running", "complete");
   private static final Set<String> AUDITED_TASK_TYPES
       = ImmutableSet.of("index", "index_parallel", "compact", "index_hadoop", "kill");
 
-  private enum TaskStateLookup
-  {
-    ALL,
-    WAITING,
-    PENDING,
-    RUNNING,
-    COMPLETE;
-
-    private static TaskStateLookup fromString(@Nullable String state)
-    {
-      if (state == null) {
-        return ALL;
-      } else {
-        return TaskStateLookup.valueOf(StringUtils.toUpperCase(state));
-      }
-    }
-  }
-
   @Inject
   public OverlordResource(
+      DruidOverlord overlord,
       TaskMaster taskMaster,
       TaskQueryTool taskQueryTool,
       IndexerMetadataStorageAdapter indexerMetadataStorageAdapter,
@@ -169,10 +139,10 @@ public class OverlordResource
       AuditManager auditManager,
       AuthorizerMapper authorizerMapper,
       WorkerTaskRunnerQueryAdapter workerTaskRunnerQueryAdapter,
-      ProvisioningStrategy provisioningStrategy,
       AuthConfig authConfig
   )
   {
+    this.overlord = overlord;
     this.taskMaster = taskMaster;
     this.taskQueryTool = taskQueryTool;
     this.indexerMetadataStorageAdapter = indexerMetadataStorageAdapter;
@@ -181,7 +151,6 @@ public class OverlordResource
     this.auditManager = auditManager;
     this.authorizerMapper = authorizerMapper;
     this.workerTaskRunnerQueryAdapter = workerTaskRunnerQueryAdapter;
-    this.provisioningStrategy = provisioningStrategy;
     this.authConfig = authConfig;
   }
 
@@ -252,7 +221,7 @@ public class OverlordResource
   @Produces(MediaType.APPLICATION_JSON)
   public Response getLeader()
   {
-    return Response.ok(taskMaster.getCurrentLeader()).build();
+    return Response.ok(overlord.getCurrentLeader()).build();
   }
 
   /**
@@ -263,7 +232,7 @@ public class OverlordResource
   @Produces(MediaType.APPLICATION_JSON)
   public Response isLeader()
   {
-    final boolean leading = taskMaster.isLeader();
+    final boolean leading = overlord.isLeader();
     final Map<String, Boolean> response = ImmutableMap.of("leader", leading);
     if (leading) {
       return Response.ok(response).build();
@@ -373,9 +342,7 @@ public class OverlordResource
                 taskInfo.getStatus().getStatusCode(),
                 RunnerTaskState.WAITING,
                 taskInfo.getStatus().getDuration(),
-                taskInfo.getStatus().getLocation() == null
-                ? TaskLocation.unknown()
-                : taskInfo.getStatus().getLocation(),
+                Configs.valueOrDefault(taskInfo.getStatus().getLocation(), TaskLocation.unknown()),
                 taskInfo.getDataSource(),
                 taskInfo.getStatus().getErrorMsg()
             )
@@ -415,14 +382,9 @@ public class OverlordResource
   {
     return asLeaderWith(
         taskMaster.getTaskQueue(),
-        new Function<TaskQueue, Response>()
-        {
-          @Override
-          public Response apply(TaskQueue taskQueue)
-          {
-            taskQueue.shutdown(taskid, "Shutdown request from user");
-            return Response.ok(ImmutableMap.of("task", taskid)).build();
-          }
+        taskQueue -> {
+          taskQueue.shutdown(taskid, "Shutdown request from user");
+          return Response.ok(ImmutableMap.of("task", taskid)).build();
         }
     );
   }
@@ -435,20 +397,15 @@ public class OverlordResource
   {
     return asLeaderWith(
         taskMaster.getTaskQueue(),
-        new Function<TaskQueue, Response>()
-        {
-          @Override
-          public Response apply(TaskQueue taskQueue)
-          {
-            final List<TaskInfo<Task, TaskStatus>> tasks = taskQueryTool.getActiveTaskInfo(dataSource);
-            if (tasks.isEmpty()) {
-              return Response.status(Status.NOT_FOUND).build();
-            } else {
-              for (final TaskInfo<Task, TaskStatus> task : tasks) {
-                taskQueue.shutdown(task.getId(), "Shutdown request from user");
-              }
-              return Response.ok(ImmutableMap.of("dataSource", dataSource)).build();
+        taskQueue -> {
+          final List<TaskInfo<Task, TaskStatus>> tasks = taskQueryTool.getActiveTaskInfo(dataSource);
+          if (tasks.isEmpty()) {
+            return Response.status(Status.NOT_FOUND).build();
+          } else {
+            for (final TaskInfo<Task, TaskStatus> task : tasks) {
+              taskQueue.shutdown(task.getId(), "Shutdown request from user");
             }
+            return Response.ok(ImmutableMap.of("dataSource", dataSource)).build();
           }
         }
     );
@@ -460,19 +417,13 @@ public class OverlordResource
   @ResourceFilters(StateResourceFilter.class)
   public Response getMultipleTaskStatuses(Set<String> taskIds)
   {
-    if (taskIds == null || taskIds.size() == 0) {
-      return Response.status(Response.Status.BAD_REQUEST).entity("No TaskIds provided.").build();
+    if (CollectionUtils.isNullOrEmpty(taskIds)) {
+      return Response.status(Response.Status.BAD_REQUEST).entity("No Task IDs provided.").build();
     }
 
-    final Optional<TaskQueue> taskQueue = taskMaster.getTaskQueue();
-    Map<String, TaskStatus> result = Maps.newHashMapWithExpectedSize(taskIds.size());
+    final Map<String, TaskStatus> result = Maps.newHashMapWithExpectedSize(taskIds.size());
     for (String taskId : taskIds) {
-      final Optional<TaskStatus> optional;
-      if (taskQueue.isPresent()) {
-        optional = taskQueue.get().getTaskStatus(taskId);
-      } else {
-        optional = taskQueryTool.getStatus(taskId);
-      }
+      final Optional<TaskStatus> optional = taskQueryTool.getTaskStatus(taskId);
       if (optional.isPresent()) {
         result.put(taskId, optional.get());
       }
@@ -487,11 +438,7 @@ public class OverlordResource
   @ResourceFilters(ConfigResourceFilter.class)
   public Response getWorkerConfig()
   {
-    if (workerConfigRef == null) {
-      workerConfigRef = configManager.watch(WorkerBehaviorConfig.CONFIG_KEY, WorkerBehaviorConfig.class);
-    }
-
-    return Response.ok(workerConfigRef.get()).build();
+    return Response.ok(taskQueryTool.getLatestWorkerConfig()).build();
   }
 
   /**
@@ -503,49 +450,11 @@ public class OverlordResource
   @ResourceFilters(ConfigResourceFilter.class)
   public Response getTotalWorkerCapacity()
   {
-    // Calculate current cluster capacity
-    Optional<TaskRunner> taskRunnerOptional = taskMaster.getTaskRunner();
-    if (!taskRunnerOptional.isPresent()) {
-      // Cannot serve call as not leader
+    if (overlord.isLeader()) {
+      return Response.ok(taskQueryTool.getTotalWorkerCapacity()).build();
+    } else {
       return Response.status(Response.Status.SERVICE_UNAVAILABLE).build();
     }
-    TaskRunner taskRunner = taskRunnerOptional.get();
-    Collection<ImmutableWorkerInfo> workers = taskRunner instanceof WorkerTaskRunner ?
-        ((WorkerTaskRunner) taskRunner).getWorkers() : ImmutableList.of();
-
-    int currentCapacity = taskRunner.getTotalCapacity();
-    int usedCapacity = taskRunner.getUsedCapacity();
-    // Calculate maximum capacity with auto scale
-    int maximumCapacity;
-    if (workerConfigRef == null) {
-      workerConfigRef = configManager.watch(WorkerBehaviorConfig.CONFIG_KEY, WorkerBehaviorConfig.class);
-    }
-    WorkerBehaviorConfig workerBehaviorConfig = workerConfigRef.get();
-    if (workerBehaviorConfig == null) {
-      // Auto scale not setup
-      log.debug("Cannot calculate maximum worker capacity as worker behavior config is not configured");
-      maximumCapacity = -1;
-    } else if (workerBehaviorConfig instanceof DefaultWorkerBehaviorConfig) {
-      DefaultWorkerBehaviorConfig defaultWorkerBehaviorConfig = (DefaultWorkerBehaviorConfig) workerBehaviorConfig;
-      if (defaultWorkerBehaviorConfig.getAutoScaler() == null) {
-        // Auto scale not setup
-        log.debug("Cannot calculate maximum worker capacity as auto scaler not configured");
-        maximumCapacity = -1;
-      } else {
-        int maxWorker = defaultWorkerBehaviorConfig.getAutoScaler().getMaxNumWorkers();
-        int expectedWorkerCapacity = provisioningStrategy.getExpectedWorkerCapacity(workers);
-        maximumCapacity = expectedWorkerCapacity == -1 ? -1 : maxWorker * expectedWorkerCapacity;
-      }
-    } else {
-      // Auto scale is not using DefaultWorkerBehaviorConfig
-      log.debug(
-          "Cannot calculate maximum worker capacity as WorkerBehaviorConfig [%s] of type [%s] does not support getting max capacity",
-          workerBehaviorConfig,
-          workerBehaviorConfig.getClass().getSimpleName()
-      );
-      maximumCapacity = -1;
-    }
-    return Response.ok(new TotalWorkerCapacityResponse(currentCapacity, maximumCapacity, usedCapacity)).build();
   }
 
   // default value is used for backwards compatibility
@@ -693,9 +602,13 @@ public class OverlordResource
     //check for valid state
     if (state != null) {
       if (!API_TASK_STATES.contains(StringUtils.toLowerCase(state))) {
+        String errorMessage = StringUtils.format(
+            "Invalid task state[%s]. Must be one of %s.",
+            state, API_TASK_STATES
+        );
         return Response.status(Status.BAD_REQUEST)
                        .type(MediaType.TEXT_PLAIN)
-                       .entity(StringUtils.format("Invalid state : %s, valid values are: %s", state, API_TASK_STATES))
+                       .entity(errorMessage)
                        .build();
       }
     }
@@ -725,7 +638,7 @@ public class OverlordResource
         taskMaster.getTaskRunner(),
         taskRunner -> {
           final List<TaskStatusPlus> authorizedList = securedTaskStatusPlus(
-              getTaskStatusPlusList(
+              taskQueryTool.getTaskStatusPlusList(
                   taskRunner,
                   TaskStateLookup.fromString(state),
                   dataSource,
@@ -739,180 +652,6 @@ public class OverlordResource
           return Response.ok(authorizedList).build();
         }
     );
-  }
-
-  private List<TaskStatusPlus> getTaskStatusPlusList(
-      TaskRunner taskRunner,
-      TaskStateLookup state,
-      @Nullable String dataSource,
-      @Nullable String createdTimeInterval,
-      @Nullable Integer maxCompletedTasks,
-      @Nullable String type
-  )
-  {
-    final Duration createdTimeDuration;
-    if (createdTimeInterval != null) {
-      final Interval theInterval = Intervals.of(StringUtils.replace(createdTimeInterval, "_", "/"));
-      createdTimeDuration = theInterval.toDuration();
-    } else {
-      createdTimeDuration = null;
-    }
-
-    // Ideally, snapshotting in taskStorage and taskRunner should be done atomically,
-    // but there is no way to do it today.
-    // Instead, we first gets a snapshot from taskStorage and then one from taskRunner.
-    // This way, we can use the snapshot from taskStorage as the source of truth for the set of tasks to process
-    // and use the snapshot from taskRunner as a reference for potential task state updates happened
-    // after the first snapshotting.
-    Stream<TaskStatusPlus> taskStatusPlusStream = getTaskStatusPlusList(
-        state,
-        dataSource,
-        createdTimeDuration,
-        maxCompletedTasks,
-        type
-    );
-    final Map<String, ? extends TaskRunnerWorkItem> runnerWorkItems = getTaskRunnerWorkItems(
-        taskRunner,
-        state,
-        dataSource,
-        type
-    );
-
-    if (state == TaskStateLookup.PENDING || state == TaskStateLookup.RUNNING) {
-      // We are interested in only those tasks which are in taskRunner.
-      taskStatusPlusStream = taskStatusPlusStream
-          .filter(statusPlus -> runnerWorkItems.containsKey(statusPlus.getId()));
-    }
-    final List<TaskStatusPlus> taskStatusPlusList = taskStatusPlusStream.collect(Collectors.toList());
-
-    // Separate complete and active tasks from taskStorage.
-    // Note that taskStorage can return only either complete tasks or active tasks per TaskLookupType.
-    final List<TaskStatusPlus> completeTaskStatusPlusList = new ArrayList<>();
-    final List<TaskStatusPlus> activeTaskStatusPlusList = new ArrayList<>();
-    for (TaskStatusPlus statusPlus : taskStatusPlusList) {
-      if (statusPlus.getStatusCode().isComplete()) {
-        completeTaskStatusPlusList.add(statusPlus);
-      } else {
-        activeTaskStatusPlusList.add(statusPlus);
-      }
-    }
-
-    final List<TaskStatusPlus> taskStatuses = new ArrayList<>(completeTaskStatusPlusList);
-
-    activeTaskStatusPlusList.forEach(statusPlus -> {
-      final TaskRunnerWorkItem runnerWorkItem = runnerWorkItems.get(statusPlus.getId());
-      if (runnerWorkItem == null) {
-        // a task is assumed to be a waiting task if it exists in taskStorage but not in taskRunner.
-        if (state == TaskStateLookup.WAITING || state == TaskStateLookup.ALL) {
-          taskStatuses.add(statusPlus);
-        }
-      } else {
-        if (state == TaskStateLookup.PENDING || state == TaskStateLookup.RUNNING || state == TaskStateLookup.ALL) {
-          taskStatuses.add(
-              new TaskStatusPlus(
-                  statusPlus.getId(),
-                  statusPlus.getGroupId(),
-                  statusPlus.getType(),
-                  statusPlus.getCreatedTime(),
-                  runnerWorkItem.getQueueInsertionTime(),
-                  statusPlus.getStatusCode(),
-                  taskRunner.getRunnerTaskState(statusPlus.getId()), // this is racy for remoteTaskRunner
-                  statusPlus.getDuration(),
-                  runnerWorkItem.getLocation(), // location in taskInfo is only updated after the task is done.
-                  statusPlus.getDataSource(),
-                  statusPlus.getErrorMsg()
-              )
-          );
-        }
-      }
-    });
-
-    return taskStatuses;
-  }
-
-  private Stream<TaskStatusPlus> getTaskStatusPlusList(
-      TaskStateLookup state,
-      @Nullable String dataSource,
-      Duration createdTimeDuration,
-      @Nullable Integer maxCompletedTasks,
-      @Nullable String type
-  )
-  {
-    final Map<TaskLookupType, TaskLookup> taskLookups;
-    switch (state) {
-      case ALL:
-        taskLookups = ImmutableMap.of(
-            TaskLookupType.ACTIVE,
-            ActiveTaskLookup.getInstance(),
-            TaskLookupType.COMPLETE,
-            CompleteTaskLookup.of(maxCompletedTasks, createdTimeDuration)
-        );
-        break;
-      case COMPLETE:
-        taskLookups = ImmutableMap.of(
-            TaskLookupType.COMPLETE,
-            CompleteTaskLookup.of(maxCompletedTasks, createdTimeDuration)
-        );
-        break;
-      case WAITING:
-      case PENDING:
-      case RUNNING:
-        taskLookups = ImmutableMap.of(
-            TaskLookupType.ACTIVE,
-            ActiveTaskLookup.getInstance()
-        );
-        break;
-      default:
-        throw new IAE("Unknown state: [%s]", state);
-    }
-
-    final Stream<TaskStatusPlus> taskStatusPlusStream = taskQueryTool.getTaskStatusPlusList(
-        taskLookups,
-        dataSource
-    ).stream();
-    if (type != null) {
-      return taskStatusPlusStream.filter(
-          statusPlus -> type.equals(statusPlus == null ? null : statusPlus.getType())
-      );
-    } else {
-      return taskStatusPlusStream;
-    }
-  }
-
-  private Map<String, ? extends TaskRunnerWorkItem> getTaskRunnerWorkItems(
-      TaskRunner taskRunner,
-      TaskStateLookup state,
-      @Nullable String dataSource,
-      @Nullable String type
-  )
-  {
-    Stream<? extends TaskRunnerWorkItem> runnerWorkItemsStream;
-    switch (state) {
-      case ALL:
-      case WAITING:
-        // waiting tasks can be found by (all tasks in taskStorage - all tasks in taskRunner)
-        runnerWorkItemsStream = taskRunner.getKnownTasks().stream();
-        break;
-      case PENDING:
-        runnerWorkItemsStream = taskRunner.getPendingTasks().stream();
-        break;
-      case RUNNING:
-        runnerWorkItemsStream = taskRunner.getRunningTasks().stream();
-        break;
-      case COMPLETE:
-        runnerWorkItemsStream = Stream.empty();
-        break;
-      default:
-        throw new IAE("Unknown state: [%s]", state);
-    }
-    if (dataSource != null) {
-      runnerWorkItemsStream = runnerWorkItemsStream.filter(item -> dataSource.equals(item.getDataSource()));
-    }
-    if (type != null) {
-      runnerWorkItemsStream = runnerWorkItemsStream.filter(item -> type.equals(item.getTaskType()));
-    }
-    return runnerWorkItemsStream
-        .collect(Collectors.toMap(TaskRunnerWorkItem::getTaskId, item -> item));
   }
 
   @DELETE
@@ -939,7 +678,7 @@ public class OverlordResource
       throw new ForbiddenException(authResult.getMessage());
     }
 
-    if (taskMaster.isLeader()) {
+    if (overlord.isLeader()) {
       try {
         final int numDeleted = indexerMetadataStorageAdapter.deletePendingSegments(dataSource, deleteInterval);
         return Response.ok().entity(ImmutableMap.of("numDeleted", numDeleted)).build();
@@ -970,23 +709,17 @@ public class OverlordResource
   {
     return asLeaderWith(
         taskMaster.getTaskRunner(),
-        new Function<TaskRunner, Response>()
-        {
-          @Override
-          public Response apply(TaskRunner taskRunner)
-          {
-            if (taskRunner instanceof WorkerTaskRunner) {
-              return Response.ok(((WorkerTaskRunner) taskRunner).getWorkers()).build();
-            } else {
-              log.debug(
-                  "Task runner [%s] of type [%s] does not support listing workers",
-                  taskRunner,
-                  taskRunner.getClass().getName()
-              );
-              return Response.serverError()
-                             .entity(ImmutableMap.of("error", "Task Runner does not support worker listing"))
-                             .build();
-            }
+        taskRunner -> {
+          if (taskRunner instanceof WorkerTaskRunner) {
+            return Response.ok(((WorkerTaskRunner) taskRunner).getWorkers()).build();
+          } else {
+            log.debug(
+                "Task runner[%s] of type[%s] does not support listing workers",
+                taskRunner, taskRunner.getClass().getName()
+            );
+            return Response.serverError()
+                           .entity(ImmutableMap.of("error", "Task Runner does not support worker listing"))
+                           .build();
           }
         }
     );
