+++ indexing-service/src/test/java/org/apache/druid/indexing/overlord/http/OverlordRedirectInfoTest.java
@@ -20,6 +20,7 @@
 package org.apache.druid.indexing.overlord.http;
 
 import com.google.common.base.Optional;
+import org.apache.druid.indexing.overlord.DruidOverlord;
 import org.apache.druid.indexing.overlord.TaskMaster;
 import org.easymock.EasyMock;
 import org.junit.Assert;
@@ -32,48 +33,48 @@ import java.net.URLEncoder;
 
 public class OverlordRedirectInfoTest
 {
-  private TaskMaster taskMaster;
+  private DruidOverlord overlord;
   private OverlordRedirectInfo redirectInfo;
 
   @Before
   public void setUp()
   {
-    taskMaster = EasyMock.createMock(TaskMaster.class);
-    redirectInfo = new OverlordRedirectInfo(taskMaster);
+    overlord = EasyMock.createMock(TaskMaster.class);
+    redirectInfo = new OverlordRedirectInfo(overlord);
   }
 
   @Test
   public void testDoLocalWhenLeading()
   {
-    EasyMock.expect(taskMaster.isLeader()).andReturn(true).anyTimes();
-    EasyMock.replay(taskMaster);
+    EasyMock.expect(overlord.isLeader()).andReturn(true).anyTimes();
+    EasyMock.replay(overlord);
     Assert.assertTrue(redirectInfo.doLocal(null));
     Assert.assertTrue(redirectInfo.doLocal("/druid/indexer/v1/leader"));
     Assert.assertTrue(redirectInfo.doLocal("/druid/indexer/v1/isLeader"));
     Assert.assertTrue(redirectInfo.doLocal("/druid/indexer/v1/other/path"));
-    EasyMock.verify(taskMaster);
+    EasyMock.verify(overlord);
   }
 
   @Test
   public void testDoLocalWhenNotLeading()
   {
-    EasyMock.expect(taskMaster.isLeader()).andReturn(false).anyTimes();
-    EasyMock.replay(taskMaster);
+    EasyMock.expect(overlord.isLeader()).andReturn(false).anyTimes();
+    EasyMock.replay(overlord);
     Assert.assertFalse(redirectInfo.doLocal(null));
     Assert.assertTrue(redirectInfo.doLocal("/druid/indexer/v1/leader"));
     Assert.assertTrue(redirectInfo.doLocal("/druid/indexer/v1/isLeader"));
     Assert.assertFalse(redirectInfo.doLocal("/druid/indexer/v1/other/path"));
-    EasyMock.verify(taskMaster);
+    EasyMock.verify(overlord);
   }
 
   @Test
   public void testGetRedirectURLWithEmptyLocation()
   {
-    EasyMock.expect(taskMaster.getRedirectLocation()).andReturn(Optional.absent()).anyTimes();
-    EasyMock.replay(taskMaster);
+    EasyMock.expect(overlord.getRedirectLocation()).andReturn(Optional.absent()).anyTimes();
+    EasyMock.replay(overlord);
     URL url = redirectInfo.getRedirectURL("query", "/request");
     Assert.assertNull(url);
-    EasyMock.verify(taskMaster);
+    EasyMock.verify(overlord);
   }
 
   @Test
@@ -82,11 +83,11 @@ public class OverlordRedirectInfoTest
     String host = "http://localhost";
     String query = "foo=bar&x=y";
     String request = "/request";
-    EasyMock.expect(taskMaster.getRedirectLocation()).andReturn(Optional.of(host)).anyTimes();
-    EasyMock.replay(taskMaster);
+    EasyMock.expect(overlord.getRedirectLocation()).andReturn(Optional.of(host)).anyTimes();
+    EasyMock.replay(overlord);
     URL url = redirectInfo.getRedirectURL(query, request);
     Assert.assertEquals("http://localhost/request?foo=bar&x=y", url.toString());
-    EasyMock.verify(taskMaster);
+    EasyMock.verify(overlord);
   }
 
   @Test
@@ -98,14 +99,14 @@ public class OverlordRedirectInfoTest
         "UTF-8"
     ) + "/status";
 
-    EasyMock.expect(taskMaster.getRedirectLocation()).andReturn(Optional.of(host)).anyTimes();
-    EasyMock.replay(taskMaster);
+    EasyMock.expect(overlord.getRedirectLocation()).andReturn(Optional.of(host)).anyTimes();
+    EasyMock.replay(overlord);
     URL url = redirectInfo.getRedirectURL(null, request);
     Assert.assertEquals(
         "http://localhost/druid/indexer/v1/task/index_hadoop_datasource_2017-07-12T07%3A43%3A01.495Z/status",
         url.toString()
     );
-    EasyMock.verify(taskMaster);
+    EasyMock.verify(overlord);
   }
 
 }
